refactor: 입출력 관련 상수 message파일로 분리
